+++ Ginger Island Mainland Adjustments/Ginger Island Mainland Adjustments/docs/changelog.md
@@ -25,6 +25,7 @@
 ### Version 1.1.3
  
 * In theory: integration with Child2NPC.
+* You can now call Pam, who will tell you if she's headed to Ginger Island.
 
 ### Version 1.1.2